% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_exist.R
\name{var_exist}
\alias{var_exist}
\title{check the variable existed in the data frame}
\usage{
var_exist(dt, var)
}
\arguments{
\item{dt}{data frame}

\item{var}{variable name}
}
\value{
value
}
\description{
check the variable existed in the data frame
}
